+++ routes/routes/user.js
@@ -17,7 +17,8 @@ module.exports = (router) => {
             cb(null, 'uploads/users/');
         },
         filename: (req, file, cb) => {
-            cb(null, file.originalname);
+            if (file)
+                cb(null, file.originalname);
         },
     });
 
@@ -35,20 +36,22 @@ module.exports = (router) => {
                 const user = new User({
                     username: username,
                     password: password,
-                    color: color.slice(0, 6),
+                    color: color.slice(0, 7),
                 });
 
-                const result = await cloudinary.v2.uploader
-                    .upload("uploads/users/" + req.file.originalname, {
-                        folder: 'chatterbox/pfps/',
-                        resource_type: 'image',
-                        public_id: user._id
+                let result = null;
+                if (req.file)
+                    result = await cloudinary.v2.uploader
+                        .upload("uploads/users/" + req.file.originalname, {
+                            folder: 'chatterbox/pfps/',
+                            resource_type: 'image',
+                            public_id: user._id
+                        });
+                const avatarURL = result !== null ? result.secure_url : result;
+                if (req.file)
+                    fs.rm("uploads/users/" + req.file.originalname, (err) => {
+                        if (err) throw err;
                     });
-                const avatarURL = result.secure_url;
-
-                fs.rm("uploads/users/" + req.file.originalname, (err) => {
-                    if (err) throw err;
-                });
                 user.avatarURL = avatarURL;
                 if (userDb && userDb.length > 0)
                     return res.status(400).send({ error: types.ErrorTypes.UNAME_NOT_AVAILABLE });
@@ -62,24 +65,28 @@ module.exports = (router) => {
     });
 
     router.get("/get_users", isAuthorized, async (req, res) => {
-        if (!req.query) return res.send({ error: types.ErrorTypes.INVALID_REQUEST });
-        const cid = req.query.cid;
-
-        const channel = await Channel.findById(cid);
-        const members = channel.members;
-        const mem2push = [];
-        for (let i = 0; i < members.length; i++) {
-            const member = await User.findById(members[i]);
-            const mem = {
-                username: member.username,
-                avatarURL: member.avatarURL,
-                status: member.status,
-                color: member.color
-            };
-            mem2push.push(mem);
+        try{
+            if (!req.query) return res.send({ error: types.ErrorTypes.INVALID_REQUEST });
+            const cid = req.query.cid;
+    
+            const channel = await Channel.findById(cid);
+            const members = channel.members;
+            const mem2push = [];
+            for (let i = 0; i < members.length; i++) {
+                const member = await User.findById(members[i]);
+                const mem = {
+                    username: member.username,
+                    avatarURL: member.avatarURL,
+                    status: member.status,
+                    color: member.color
+                };
+                mem2push.push(mem);
+            }
+    
+            return res.status(200).send({ members: mem2push });
+        }catch(e){
+            return res.status(518).send({error: types.ErrorTypes.UNKNOWN_ERROR});
         }
-
-        return res.status(200).send({ members: mem2push });
     })
 
     router.get("/search_user", isAuthorized, async (req, res) => {
